requirements.txt
flask>=3.0.0
twilio>=9.0.0
openai>=1.30.0
python-dotenv>=1.0.0
